Node Deleting
+ Cant delete nodes accurately :/

+++ BinaryTrees-raylib/RayGUI_Widgets/BinaryTreeMain.cpp
@@ -7,18 +7,14 @@ BinaryTree::BinaryTree() {
 }
 
 BinaryTree::~BinaryTree() {
-	while (mRoot_ptr) {
-		DeleteNode(mRoot_ptr->NodeData());
-	}
+	delete mRoot_ptr;
 }
 
  /// INSERT NODE
 /* Insert node from tree */
 void BinaryTree::InsertNode(int newData) {
-	TreeNode* newNode = new TreeNode(newData, nullptr, nullptr, nullptr);
-
 	if (mRoot_ptr == nullptr) { //If a rootNode doesnt exist
-		mRoot_ptr = newNode;   //Create one
+		mRoot_ptr = new TreeNode(newData, nullptr, nullptr, nullptr); //Create one
 		return;
 	}
 	else {
@@ -32,8 +28,12 @@ void BinaryTree::DeleteNode(int mNodeToDelete) {
 	if (mRoot_ptr == nullptr) { return; }
 
 	TreeNode* nodeToDelete = Get(mNodeToDelete);
-
-	mRoot_ptr->DeleteNode(nodeToDelete);
+	if (nodeToDelete != mRoot_ptr)
+		mRoot_ptr->DeleteNode(nodeToDelete);
+	else {
+		delete mRoot_ptr;
+		mRoot_ptr == nullptr;
+	}
 }
 
  /// FIND NODE
@@ -43,47 +43,48 @@ TreeNode* BinaryTree::Get(int dataToFind) { //Deleteing child nodes gets stick i
 	TreeNode* parentNode = nullptr;
 
 	while (currentNode != nullptr) {
-		if (dataToFind == currentNode->NodeData()) { return currentNode, parentNode; }
+		if (dataToFind == currentNode->NodeData()) { return currentNode; }
 
 		if (dataToFind < currentNode->NodeData()) {
-			currentNode->LeftNode();
+			currentNode = currentNode->LeftNode();
 		}
 		else {
-			currentNode->RightNode(); 
+			currentNode = currentNode->RightNode();
 		}
 	}
-
 	return FindNode(dataToFind, currentNode, parentNode) ? currentNode : nullptr;
 }
 
-bool BinaryTree::FindNode(int dataToSearch, TreeNode*& outNode_dblptr, TreeNode*& OutParent_dblptr) {
+bool BinaryTree::FindNode(int dataToSearch, TreeNode*& outNode, TreeNode*& outParent) {
 	return false;
 }
 
  /// DRAW
 /* Draw the binary trees root node */
 void BinaryTree::DrawRoot(TreeNode* selectedNode) {
+	if (mRoot_ptr == nullptr) { return; }
+
 	DrawTree(mRoot_ptr, GetScreenWidth() / 2, 40, 400, selectedNode);
 }
 
 /* Draw the binary trees nodes */
-void BinaryTree::DrawTree(TreeNode* currentNode, int posX, int posY, int horizontalSpacing, TreeNode* selectedNode) {
-	if (mRoot_ptr == nullptr) { return; }
-	horizontalSpacing /= 2;
+void BinaryTree::DrawTree(TreeNode* currentNode, int posX, int posY, int nodeSeperation, TreeNode* selectedNode) {
+	if (mRoot_ptr == nullptr || currentNode == nullptr) { return; }
 
-	if (currentNode) {
-		if (currentNode->BranchesLeft()) {
-			DrawLine(posX, posY, (posX - horizontalSpacing), (posY + 80), BROWN);
+	TreeNode* nodeToDraw = currentNode; //Just making it easier to see
+	nodeSeperation /= 2;
 
-			DrawTree(currentNode->LeftNode(), posX - horizontalSpacing, posY + 80, horizontalSpacing, selectedNode);
+	if (nodeToDraw) {
+		if (nodeToDraw->BranchesLeft()) {
+			DrawLine(posX, posY, (posX - nodeSeperation), (posY + 80), BROWN);
+			DrawTree(nodeToDraw->LeftNode(), (posX - nodeSeperation), (posY + 80), nodeSeperation, selectedNode);
 		}
 
-		if (currentNode->BranchesRight()) {
-			DrawLine(posX, posY, (posX + horizontalSpacing), (posY + 80), BEIGE);
-
-			DrawTree(currentNode->RightNode(), posX + horizontalSpacing, posY + 80, horizontalSpacing, selectedNode);
+		if (nodeToDraw->BranchesRight()) {
+			DrawLine(posX, posY, (posX + nodeSeperation), (posY + 80), BEIGE);
+			DrawTree(nodeToDraw->RightNode(), (posX + nodeSeperation), (posY + 80), nodeSeperation, selectedNode);
 		}
 
-		currentNode->DrawNode(posX, posY, (selectedNode == currentNode));
+		nodeToDraw->DrawNode(posX, posY);
 	}
 }
